Add unbound uncertainty/durations for tasks
For a task named `Task`, unbound uncertainty with a lower bound of `x` can be modeled with a duration child feature named `dTask >= x`.

Duration constraints cannot be created from an unbound durations feature.
If duration constraints in the range of the unbound uncertainty are needed, a separate duration feature needs to be created.